+++ elastic/tox.ini
@@ -1,14 +1,14 @@
 [tox]
 minversion = 2.0
 basepython = py27
-#envlist = elastic{0.90,1.0,1.1,1.2,2.4,5.2,5.3,5.4,5.5,5.6,6.0,6.1,6.2}, flake8
-envlist = elastic-bench
+envlist = elastic{0.90,1.0,1.1,1.2,2.4,5.2,5.3,5.4,5.5,5.6,6.0,6.1,6.2}, bench, flake8
+
 [testenv]
 platform = linux|darwin|win32
-deps = 
+deps =
     ../datadog_checks_base
     -rrequirements-dev.txt
-commands = 
+commands =
     pip install --require-hashes -r requirements.txt
     pytest -v --benchmark-skip
 passenv =
@@ -18,72 +18,72 @@ passenv =
 # Supported Versions
 #
 [testenv:elastic0.90]
-setenv = 
+setenv =
     ELASTIC_DOCKER=datadog/docker-library:elasticsearch_
     ELASTIC_VERSION=0_90
 
 [testenv:elastic1.0]
-setenv = 
+setenv =
     ELASTIC_DOCKER=datadog/docker-library:elasticsearch_
     ELASTIC_VERSION=1_0
 
 [testenv:elastic1.1]
-setenv = 
+setenv =
     ELASTIC_DOCKER=datadog/docker-library:elasticsearch_
     ELASTIC_VERSION=1_1
 
 [testenv:elastic1.2]
-setenv = 
+setenv =
     ELASTIC_DOCKER=datadog/docker-library:elasticsearch_
     ELASTIC_VERSION=1_2
 
 [testenv:elastic2.4]
-setenv = 
+setenv =
     ELASTIC_DOCKER=elasticsearch:
     ELASTIC_VERSION=2.4.6
 
 [testenv:elastic5.2]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=5.2.1
 
 [testenv:elastic5.3]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=5.3.3
 
 [testenv:elastic5.4]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=5.4.3
 
 [testenv:elastic5.5]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=5.5.3
 
 [testenv:elastic5.6]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=5.6.9
 
 [testenv:elastic6.0]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=6.0.1
 
 [testenv:elastic6.1]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=6.1.4
 
 [testenv:elastic6.2]
-setenv = 
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=6.2.4
 
-[testenv:elastic-bench]
-setenv = 
+[testenv:bench]
+setenv =
     ELASTIC_DOCKER=docker.elastic.co/elasticsearch/elasticsearch:
     ELASTIC_VERSION=6.2.4
 commands =
